Changed the annoying issue with layers being made over layers on the map.
There was a problem where whenever a new layer was added, the existing
layer would remain underneath. Now, all existing layers are removed
prior to creating new layers. This may be faster, although now that I am
thinking about, I wonder if the fastest way to work with this page would
be to have an index of the most common layers loaded to reduce overall
time for changes to the filters...I am not sure, but this seems
possible...